requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv-contrib-python==4.5.1.48
Pillow
scipy